refactor(equivalence): reroute pages
