+++ pages/tools/equivalent-value/index.tsx
@@ -15,41 +15,18 @@ import { EquivalenceToolbar } from '../../../components/ui/EquivalenceToolbar';
 import { AddPaymentButton } from '../../../components/ui/AddPaymentButton';
 import { authOptions } from '../../api/auth/[...nextauth]';
 
-const EquivalentValuePage: NextPage = () => {
+const EquivalentValueGroupsPage: NextPage = () => {
   return (
     <EquivalentValueProvider>
       <BasePageLayout
-        title="Equivalent Value Calculator"
-        description="Calculate the equivalent value of a set of payments in a given time."
-        displayFooter={false}
-        sx={{
-          display: undefined,
-          position: 'relative',
-        }}
+        title="Equivalent Value Calculation - Groups"
+        description="Payment groups"
       >
-        <Box
-          sx={{
-            width: '100%',
-            height: '100%',
-            position: 'absolute',
-          }}
-        >
-          <EquivalentValuePageContent />
-        </Box>
-        <Box
-          sx={{
-            position: 'absolute',
-            bottom: 15,
-            right: 15,
-          }}
-        >
-          <AddPaymentButton />
-        </Box>
       </BasePageLayout>
     </EquivalentValueProvider>
   );
 }
-export default EquivalentValuePage;
+export default EquivalentValueGroupsPage;
 
 export const getServerSideProps: GetServerSideProps = async (context: GetServerSidePropsContext) => {
   const session = await unstable_getServerSession(context.req, context.res, authOptions);
@@ -66,167 +43,3 @@ export const getServerSideProps: GetServerSideProps = async (context: GetServerS
   };
 }
 
-const EquivalentValuePageContent: FC = () => {
-  const { group: { payments } } = useEquivalentValueContext();
-  return (
-    <Stack
-      sx={{
-        width: '100%',
-        height: '100%',
-      }}
-    >
-      <EquivalenceToolbar />
-      {
-        payments.length === 0
-          ? <NoPaymentsMessage />
-          : <EquivalentValueChart />
-      }
-    </Stack>
-  );
-}
-
-const NoPaymentsMessage: FC = () => {
-  return (
-    <Box
-      sx={{
-        width: '100%',
-        height: '100%',
-        display: 'flex',
-      }}
-    >
-      <Typography
-        textAlign="center"
-        sx={{
-          m: 'auto',
-          typography: {
-            xs: 'h5',
-            sm: 'h4',
-          },
-        }}
-      >
-        No payments
-      </Typography>
-    </Box>
-  );
-}
-
-const EquivalentValueChart: FC = () => {
-  const [blockWidth, setBlockWidth] = useState<number>(50);
-  const [blockHeight, setBlockHeight] = useState<number>(50);
-  const { group } = useEquivalentValueContext();
-  const maxPosition = useMemo(
-    () => Math.max(...group.payments.map<number>(p => {
-      switch (p.type) {
-        case IPaymentType.single:
-          return p.position;
-        case IPaymentType.uniformSeries:
-          return p.position + p.periods;
-      }
-    })),
-    [group.payments],
-  );
-
-  const periodsBar = (
-    <PeriodsBar
-      positionsCount={maxPosition + 10}
-      blockWidth={blockWidth}
-    />
-  );
-
-  return (
-    <Box
-      sx={{
-        flex: 1,
-        position: 'relative',
-        overflowX: 'auto',
-        px: 3,
-      }}
-    >
-      <Stack
-        direction="row"
-        sx={{
-          position: 'absolute',
-          height: '100%',
-          overflowX: 'auto',
-        }}
-      >
-        {
-          Array.from(
-            { length: maxPosition + 10 },
-            (x, i) => (
-              <Box
-                key={i}
-                sx={{
-                  width: blockWidth,
-                  borderLeft: '.2px solid white',
-                  opacity: .25,
-                }}
-              >
-              </Box>
-            ),
-          )
-        }
-      </Stack>
-      <Stack
-        sx={{
-          position: 'absolute',
-          height: '100%',
-        }}
-      >
-        {periodsBar}
-        <Box
-          sx={{
-            position: 'relative',
-            flex: 1,
-            overflowY: 'auto',
-            py: 2,
-          }}
-        >
-          <Stack
-            spacing={2}
-            sx={{
-              position: 'absolute',
-            }}
-          >
-            {
-              group.payments.map<JSX.Element>((p, i) => {
-                return (
-                  <Box
-                    key={p.name}
-                    sx={{
-                      height: blockHeight,
-                    }}
-                  >
-                    {
-                      (() => {
-                        switch (p.type) {
-                          case IPaymentType.single: {
-                            return <SinglePayment
-                              key={p.name}
-                              payment={p}
-                              blockWidth={blockWidth}
-                              blockHeight={blockHeight}
-                            />;
-                          }
-                          case IPaymentType.uniformSeries: {
-                            return <UniformSeriesPayment
-                              key={p.name}
-                              payment={p}
-                              blockWidth={blockWidth}
-                              blockHeight={blockHeight}
-                            />;
-                          }
-                        }
-                      })()
-                    }
-                  </Box>
-                )
-              })
-            }
-          </Stack>
-        </Box>
-        {periodsBar}
-      </Stack>
-    </Box>
-  );
-}
